% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_use.R
\name{checkDependentPackageUsage}
\alias{checkDependentPackageUsage}
\title{Check Package Use in Code}
\usage{
checkDependentPackageUsage(package_name, code)
}
\arguments{
\item{package_name}{Name of the package to check against}

\item{code}{A character vector of code chunks to check for package use}
}
\value{
An object of class \code{TODO}
}
\description{
This will check what functions contained within a specified package are used in code.
}
