% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{checkPackageDependencyUse}
\alias{checkPackageDependencyUse}
\title{Check Package Dependencies}
\usage{
checkPackageDependencyUse(path = ".", include_suggests = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{Path to the package root directory.}

\item{include_suggests}{Logical, should the "Suggests" field also be checked for package dependencies or just
the Depends and Imports?}

\item{verbose}{Logical, should informative messages be printed during the dependency evaluation?}
}
\value{
An object of class \code{multi_package_usage}, a named list of the dependencies used, each containing a
\code{data.frame} of all the functions within the package and how often they are used within the project.

The result will flag any packages that have been rarely used in the project.
}
\description{
`checkPackageDependencyUse()` checks the \code{DESCRIPTION} file for packages mentioned in the \code{Depends}
and \code{Imports} fields.

Using these packages, it then checks the scripts in the R directory to check for function use of each package.
}
\examples{
\dontrun{
dependency_use <- checkPackageDependencyUse()
summary(dependency_use)
}

}
\seealso{
\code{\link{checkProjectDependencyUse}}, \code{\link{checkShinyDependencyUse}}
}
