% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_use.R
\name{checkPackagesUsage}
\alias{checkPackagesUsage}
\alias{checkPackageUsage}
\title{Check Package Use in Code}
\usage{
checkPackagesUsage(package_names, code, verbose = TRUE)

checkPackageUsage(package_name, code)
}
\arguments{
\item{package_names, package_name}{Name(s) of the package to check against}

\item{code}{A character vector of code chunks to check for package use}

\item{verbose}{Logical, should informative messages be printed during the dependency evaluation?}
}
\value{
\code{checkPackageUsage} will return a data.frame of class \code{package_usage}. When printed it will show
a summary of the package usage.

\code{checkPackagesUsage} will return a list of class \code{multi_package_usage}. When printed it will show
a summary of all packages mentioned, and flag any rarely used.
}
\description{
This will check what functions contained within a specified package are used in code.
}
