% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_use.R
\name{checkPackagesUse}
\alias{checkPackagesUse}
\alias{checkPackageUse}
\title{Check Package Use in Code}
\usage{
checkPackagesUse(package_name, code)

checkPackageUse(package_name, code)
}
\arguments{
\item{package_name}{Name(s) of the package to check against}

\item{code}{A character vector of code chunks to check for package use}
}
\value{
\code{checkPackageUse} will return a data.frame of class \code{function_usage}. When printed it will show
a summary of the package usage.

\code{checkPackagesUse} will return a list of class \code{package_usage}. When printed it will show
a summary of all packages mentioned, and flag any rarely used.
}
\description{
This will check what functions contained within a specified package are used in code.
}
