% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_check.R
\name{checkProjectDependencyUse}
\alias{checkProjectDependencyUse}
\title{Check Project Package Dependencies}
\usage{
checkProjectDependencyUse(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{Path to the directories to search for R scripts. By default looks in the current working directory.}

\item{recursive}{Logical. Should the R file search recurse into directories?}
}
\description{
This...
}
\examples{
\dontrun{
dependency_use <- checkProjectDependencyUse()
summary(dependency_use)
}

}
\seealso{
\code{\link{checkPackageDependencyUse}}, \code{\link{checkShinyDependencyUse}}
}
