% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_check.R
\name{checkShinyDependencies}
\alias{checkShinyDependencies}
\title{Check Shiny Application Dependencies}
\usage{
checkShinyDependencies(path = ".", r_scripts_dir = "R")
}
\arguments{
\item{path}{Path to the root application directory}

\item{r_scripts_dir}{Subdirectories in the shiny application that contain R scripts, write as relative paths to
\code{path}. Default is set to \code{R}}
}
\description{
This...
}
\examples{
\dontrun{
checkShinyDependencies()
}

}
