% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_check.R
\name{checkShinyDependencyUse}
\alias{checkShinyDependencyUse}
\title{Check Shiny Application Dependencies}
\usage{
checkShinyDependencyUse(path = ".", r_scripts_dir = "R")
}
\arguments{
\item{path}{Path to the application root directory}

\item{r_scripts_dir}{Subdirectories in the shiny application that contain R scripts, write as relative paths to
\code{path}. Default is set to \code{R}}
}
\description{
This...
}
\examples{
\dontrun{
dependency_use <- checkShinyDependencyUse()
summary(dependency_use)
}

}
\seealso{
\code{\link{checkProjectDependencyUse}}, \code{\link{checkPackageDependencyUse}}
}
