% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_package_use.R
\name{extractPackageCalls}
\alias{extractPackageCalls}
\title{Extract Package Calls}
\usage{
extractPackageCalls(code)
}
\arguments{
\item{code}{A character vector of code chunks to check for package calls}
}
\description{
The ability to search for package loading (via \code{library()} or \code{requireNamespace}), or by direct
referencing through \code{"pkg::function"}
}
