% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_use.R
\name{getPackageFunctions}
\alias{getPackageFunctions}
\alias{getInternalPackageFunctions}
\title{Get Package Functions}
\usage{
getPackageFunctions(package_name)

getInternalPackageFunctions(package_name)
}
\arguments{
\item{package_name}{Name of the package to find function names for}
}
\description{
In order to check if packages are included, need to extract all functions from the package.
}
