% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package_use.R
\name{print.package_usage}
\alias{print.package_usage}
\title{Package Usage Printing Options}
\usage{
\method{print}{package_usage}(
  x,
  ...,
  warn_percent_usage = 0.2,
  warn_number_usage = 3,
  include_used_functions = TRUE
)
}
\arguments{
\item{x}{A \code{package_usage} data.frame}

\item{warn_percent_usage}{Minimum percent of functions to be used within a dependent package. Default is \code{20\%}}

\item{warn_number_usage}{Minimum number of functions to be used within a dependent package. Default is \code{3}}

\item{include_used_functions}{Logical, should the functions that are used be included? Default is \code{TRUE}}
}
\description{
Package Usage Printing Options
}
\details{
Package usage must be below both thresholds for the warning to appear. With the defaults, if a package has fewer than
5 functions then only 1 function is required to prevent a warning message to appear.
}
