% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{readPackageRFiles}
\alias{readPackageRFiles}
\alias{readDirectoryRFiles}
\alias{readRFiles}
\title{Read R Files}
\usage{
readPackageRFiles(path = ".")

readDirectoryRFiles(path = ".")

readRFiles(files, path = NULL)
}
\arguments{
\item{path}{Location of the package or directory to import R files from}

\item{files}{Vector of R file paths to import}
}
\value{
A character vector containing unique code chunks in the specified directory/files.
}
\description{
Reading R files for package dependency checks.

\code{readPackageFiles} assumes the selected path is a package, and will check the \code{R} subdirectory for files
to read. \code{readDirectoryFiles} is a more generic version of \code{readPackageFiles}
}
\details{
The default directory for \code{readFiles} is \code{NULL}, to allow the reading of files from multiple directories.
}
\examples{
\dontrun{
readPackageRFiles()

# When running outside of a package
readDirectoryRFiles("R")
}

}
