% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.multi_package_usage}
\alias{summary.multi_package_usage}
\title{Package Usage Summary}
\usage{
\method{summary}{multi_package_usage}(
  object,
  warn_percent_usage = 0.2,
  warn_number_usage = 3,
  ignore_low_usage_packages = character(),
  ...
)
}
\arguments{
\item{object}{A \code{package_usage} data.frame}

\item{warn_percent_usage}{Minimum percent of functions to be used within a dependent package. Default is \code{20\%}}

\item{warn_number_usage}{Minimum number of functions to be used within a dependent package. Default is \code{3}}

\item{ignore_low_usage_packages}{A vector of packages to ignore the low usage of, usually when already aware of the
low usage, but the dependent package is necessary for the project.}

\item{...}{Not used}
}
\description{
Package Usage Summary
}
\details{
Package usage must be below both thresholds for the warning to appear. With the defaults, if a package has fewer than
5 functions then only 1 function is required to prevent a warning message to appear.
}
