% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_use.R
\name{checkFunctionUse}
\alias{checkFunctionUse}
\title{Check Function Usage in Code}
\usage{
checkFunctionUse(function_name, code, package_name = NULL)
}
\arguments{
\item{function_name}{Name of the function to check against}

\item{code}{A character vector of code chunks to check for function use}

\item{package_name}{Name of the package that \code{function_name} is contained in}
}
\description{
Counting the frequency that a function is called within selected code.
}
\details{
If \code{package_name} is left as \code{NULL}, then there is a chance the checks will find times
where the function is used, but is explicitly called from another package.

There are three stages of checking:
- If the function name exists anywhere in the code chunks
- If it has not been previously assigned in a code chunk
- If the function name is not quoted

At each stage, only keep chunks where the function has been called to improve speed
}
