% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{checkPackageDependencies}
\alias{checkPackageDependencies}
\title{Check Package Dependencies}
\usage{
checkPackageDependencies(path = ".", include_suggests = FALSE)
}
\arguments{
\item{path}{Path to the root package directory}

\item{include_suggests}{Logical, should the "Suggests" field also be checked for package dependencies or just
the Depends and Imports?}
}
\description{
This...
}
\examples{
\dontrun{
checkPackageDependencies()
}

}
