% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{checkPackageDependencyUse}
\alias{checkPackageDependencyUse}
\title{Check Package Dependencies}
\usage{
checkPackageDependencyUse(path = ".", include_suggests = FALSE)
}
\arguments{
\item{path}{Path to the package root directory.}

\item{include_suggests}{Logical, should the "Suggests" field also be checked for package dependencies or just
the Depends and Imports?}
}
\description{
This...
}
\examples{
\dontrun{
dependency_use <- checkPackageDependencyUse()
summary(dependency_use)
}

}
\seealso{
\code{\link{checkProjectDependencyUse}}, \code{\link{checkShinyDependencyUse}}
}
